+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpExporterOptions.cs
@@ -44,16 +44,19 @@ namespace OpenTelemetry.Exporter
 
         internal readonly Func<HttpClient> DefaultHttpClientFactory;
 
+        private const string DefaultGrpcEndpoint = "http://localhost:4317";
+        private const string DefaultHttpProtobufEndpoint = "http://localhost:4318";
+        private const OtlpExportProtocol DefaultOtlpExportProtocol = OtlpExportProtocol.Grpc;
+
+        private OtlpExportProtocol protocol;
+        private Uri endpoint;
+        private bool isCustomEndpointSet;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="OtlpExporterOptions"/> class.
         /// </summary>
         public OtlpExporterOptions()
         {
-            if (EnvironmentVariableHelper.LoadUri(EndpointEnvVarName, out Uri endpoint))
-            {
-                this.Endpoint = endpoint;
-            }
-
             if (EnvironmentVariableHelper.LoadString(HeadersEnvVarName, out string headersEnvVar))
             {
                 this.Headers = headersEnvVar;
@@ -66,20 +69,29 @@ namespace OpenTelemetry.Exporter
 
             if (EnvironmentVariableHelper.LoadString(ProtocolEnvVarName, out string protocolEnvVar))
             {
-                var protocol = protocolEnvVar.ToOtlpExportProtocol();
-                if (protocol.HasValue)
+                var parsedProtocol = protocolEnvVar.ToOtlpExportProtocol();
+                if (parsedProtocol.HasValue)
                 {
-                    this.Protocol = protocol.Value;
+                    this.Protocol = parsedProtocol.Value;
                 }
                 else
                 {
                     throw new FormatException($"{ProtocolEnvVarName} environment variable has an invalid value: '${protocolEnvVar}'");
                 }
             }
+            else
+            {
+                this.Protocol = DefaultOtlpExportProtocol;
+            }
+
+            if (EnvironmentVariableHelper.LoadUri(EndpointEnvVarName, out Uri parsedEndpoint))
+            {
+                this.SetEndpoint(parsedEndpoint, custom: true);
+            }
 
             this.HttpClientFactory = this.DefaultHttpClientFactory = () =>
             {
-                return new HttpClient()
+                return new HttpClient
                 {
                     Timeout = TimeSpan.FromMilliseconds(this.TimeoutMilliseconds),
                 };
@@ -89,9 +101,19 @@ namespace OpenTelemetry.Exporter
         /// <summary>
         /// Gets or sets the target to which the exporter is going to send telemetry.
         /// Must be a valid Uri with scheme (http or https) and host, and
-        /// may contain a port and path. The default value is http://localhost:4317.
+        /// may contain a port and path. The default value is
+        /// * http://localhost:4317 for <see cref="OtlpExportProtocol.Grpc"/>
+        /// * http://localhost:4318 for <see cref="OtlpExportProtocol.HttpProtobuf"/>.
         /// </summary>
-        public Uri Endpoint { get; set; } = new Uri("http://localhost:4317");
+        public Uri Endpoint
+        {
+            get => this.endpoint;
+            set
+            {
+                this.SetEndpoint(value, custom: true);
+                this.ProgrammaticallyModifiedEndpoint = true;
+            }
+        }
 
         /// <summary>
         /// Gets or sets optional headers for the connection. Refer to the <a href="https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/protocol/exporter.md#specifying-headers-via-environment-variables">
@@ -107,7 +129,24 @@ namespace OpenTelemetry.Exporter
         /// <summary>
         /// Gets or sets the the OTLP transport protocol. Supported values: Grpc and HttpProtobuf.
         /// </summary>
-        public OtlpExportProtocol Protocol { get; set; } = OtlpExportProtocol.Grpc;
+        public OtlpExportProtocol Protocol
+        {
+            get => this.protocol;
+            set
+            {
+                this.protocol = value;
+
+                switch (value)
+                {
+                    case OtlpExportProtocol.Grpc:
+                        this.SetEndpoint(new Uri(DefaultGrpcEndpoint));
+                        break;
+                    case OtlpExportProtocol.HttpProtobuf:
+                        this.SetEndpoint(new Uri(DefaultHttpProtobufEndpoint));
+                        break;
+                }
+            }
+        }
 
         /// <summary>
         /// Gets or sets the export processor type to be used with the OpenTelemetry Protocol Exporter. The default value is <see cref="ExportProcessorType.Batch"/>.
@@ -167,5 +206,26 @@ namespace OpenTelemetry.Exporter
         /// </list>
         /// </remarks>
         public Func<HttpClient> HttpClientFactory { get; set; }
+
+        /// <summary>
+        /// Gets a value indicating whether <see cref="Endpoint" /> was programmatically modified.
+        /// </summary>
+        internal bool ProgrammaticallyModifiedEndpoint { get; private set; }
+
+        private void SetEndpoint(Uri value, bool custom = false)
+        {
+            if (custom)
+            {
+                this.isCustomEndpointSet = true;
+                this.endpoint = value;
+            }
+            else
+            {
+                if (!this.isCustomEndpointSet)
+                {
+                    this.endpoint = value;
+                }
+            }
+        }
     }
 }